% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacciseq.R
\name{generate_fibonacci}
\alias{generate_fibonacci}
\title{Generate Fibonacci Sequence}
\usage{
generate_fibonacci(n, start_values = c(0, 1))
}
\arguments{
\item{n}{The number of terms in the Fibonacci sequence to generate.
This parameter is named \code{n} as it is a conventional way to denote the number of terms in sequences and iterations.}

\item{start_values}{A numeric vector of length 2 to specify the first two starting values of the sequence.
The name \code{start_values} was chosen to clearly indicate its role in initializing the sequence.}
}
\value{
A numeric vector containing the first \code{n} terms of the Fibonacci sequence.
}
\description{
Generates a Fibonacci sequence with a specified number of terms. The sequence starts with two given starting values and each subsequent term is the sum of the previous two.
}
\examples{
generate_fibonacci(n = 10)
generate_fibonacci(n = 10, start_values = c(2, 3))

}
